backend/requirements.txt
django==5.1
django-ninja==1.3.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
uvicorn==0.34.0
requests==2.32.3
django-ratelimit==4.1.0
celery==5.4.0
pillow==11.1.0
django-allauth==65.3.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
python-docx==1.1.2